+++ meow-helpers.el
@@ -41,7 +41,7 @@ Optional argument ARGS key definitions."
   (mapcar (lambda (key-def)
             (define-key meow-leader-keymap
               (kbd (car key-def))
-              (meow--parse-key-def (cdr key-def))))
+              (meow--parse-def (cdr key-def))))
           args))
 
 (defun meow-normal-define-key (&rest args)
@@ -54,7 +54,7 @@ Optional argument ARGS key definitions."
   (mapcar (lambda (key-def)
             (define-key meow-normal-state-keymap
               (kbd (car key-def))
-              (meow--parse-key-def (cdr key-def))))
+              (meow--parse-def (cdr key-def))))
           args))
 
 (defun meow-motion-overwrite-define-key (&rest args)
@@ -62,7 +62,7 @@ Optional argument ARGS key definitions."
   (mapc (lambda (key-def)
           (define-key meow-motion-state-keymap
             (kbd (car key-def))
-            (meow--parse-key-def (cdr key-def))))
+            (meow--parse-def (cdr key-def))))
         args)
   (cl-loop for arg in args do
            (add-to-list 'meow--motion-overwrite-keys (car arg))))