+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 ### Enhancements
 * Remove paredit shims, no longer needed.
 * [#110](https://github.com/meow-edit/meow/issues/110) Only disable hint overlay for modes in `meow-expand-exclude-mode-list`.
+* Add custom variable `meow-motion-remap-prefix.`
 
 ## 1.1.1 (2021-12-06)